Update ps-input-time.js
An uptade fixing some bugs: 
restrict the hour and minute inputs to deny the user to write more than 2 numbers, e.g.: 345:56PM not allowed. 
To deny the user to write letters on hour and minutes part, e.g.: tre:TrhPM not allowed. 
To control the input values e.g.: 35:98PM not allowed.